refactor some code into its own file
